Fixed the data types of the local variables used in the calls to ioctlsocket() to conform to the prototypes in order to clean up a type mismatch problem during compilation.
git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@62105 13f79535-47bb-0310-9956-ffa450edef68

+++ network_io/win32/sockopt.c
@@ -60,7 +60,7 @@
 
 apr_status_t soblock(SOCKET sd)
 {
-    int zero = 0;
+    u_long zero = 0;
 
     if (ioctlsocket(sd, FIONBIO, &zero) == SOCKET_ERROR) {
         return apr_get_netos_error();
@@ -70,7 +70,7 @@ apr_status_t soblock(SOCKET sd)
 
 apr_status_t sononblock(SOCKET sd)
 {
-    int one = 1;
+    u_long one = 1;
 
     if (ioctlsocket(sd, FIONBIO, &one) == SOCKET_ERROR) {
         return apr_get_netos_error();
